extra verbose and split tests
