+++ .github/workflows/python-package.yml
@@ -11,7 +11,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        files:  # files below + test_workshoplinks.py + test_external.py = 36 total number of test files
+        files:  # files below + test_workshoplinks.py + test_external.py = 37 total number of test files
           - "lmfdb/abvar/fq/test_av.py lmfdb/abvar/fq/test_browse_page.py lmfdb/number_fields/test_numberfield.py lmfdb/bianchi_modular_forms/test_bmf.py"
           - "lmfdb/galois_groups/test_galoisgroup.py lmfdb/siegel_modular_forms/test_siegel_modular_forms.py"
           - "lmfdb/tests/test_dynamic_knowls.py lmfdb/tests/test_root.py lmfdb/hecke_algebras/test_hecke_algebras.py lmfdb/tests/test_homepage.py lmfdb/elliptic_curves/test_ell_curves.py lmfdb/elliptic_curves/test_browse_page.py"
@@ -19,7 +19,7 @@ jobs:
           - "lmfdb/cluster_pictures/test_clusterpicture.py lmfdb/local_fields/test_localfields.py lmfdb/ecnf/test_ecnf.py lmfdb/ecnf/test_isog_class.py lmfdb/api/test_api.py lmfdb/characters/test_characters.py"
           - "lmfdb/users/test_users.py lmfdb/lattice/test_lattice.py lmfdb/maass_forms/test_maass.py lmfdb/higher_genus_w_automorphisms/test_hgcwa.py lmfdb/belyi/test_belyi.py lmfdb/tests/test_utils.py"
           - "lmfdb/artin_representations/test_artin_representation.py lmfdb/genus2_curves/test_genus2_curves.py"
-          - "lmfdb/classical_modular_forms/test_cmf.py"
+          - "lmfdb/classical_modular_forms/test_cmf.py lmfdb/classical_modular_forms/test_cmf2.py"
           - "lmfdb/lfunctions/test_lfunctions.py"
 
     steps:
@@ -81,7 +81,7 @@ jobs:
     - name: checking that we didn't miss any test files
       shell: bash -l {0}
       # If this fails you need to update the file list above and file count
-      run: test $(find lmfdb -name 'test_*.py'  | wc -l) -eq 36
+      run: test $(find lmfdb -name 'test_*.py'  | wc -l) -eq 37
 
     - name: Run LMFDB tests and linting
       shell: bash -l {0}
