+++ tox.ini
@@ -20,4 +20,4 @@ passenv =
 commands =
     pyflakes
     pylint --py3k --score=no -d W1618,W1619,W1633,W1653 lmfdb/
-    parallel sage -python -m pytest ::: {posargs}
+    parallel sage -python -m pytest -vv ::: {posargs}
